+++ gradle-plugin/src/test/kotlin/software/ralf/app/platform/gradle/AppPlatformPluginDependencyTest.kt
@@ -71,6 +71,7 @@ class AppPlatformPluginDependencyTest {
     val project = createProject(name = "impl")
     project.plugins.apply(PluginIds.KOTLIN_MULTIPLATFORM)
     project.plugins.apply(AppPlatformPlugin::class.java)
+
     project.appPlatform.enableMetro(true)
     project.appPlatform.addImplModuleDependencies(true)