+++ metro-extensions/contribute/impl-compiler-plugin/src/main/kotlin/software/ralf/app/platform/metro/compiler/AppPlatformMetroExtensionsPluginComponentRegistrar.kt
@@ -5,6 +5,9 @@ import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
 import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrarAdapter
+import software.ralf.app.platform.metro.compiler.renderer.ContributesRendererIrExtension
+import software.ralf.app.platform.metro.compiler.robot.ContributesRobotIrExtension
+import software.ralf.app.platform.metro.compiler.scoped.ContributesScopedIrExtension
 
 @AutoService(CompilerPluginRegistrar::class)
 public class AppPlatformMetroExtensionsPluginComponentRegistrar : CompilerPluginRegistrar() {
@@ -13,11 +16,8 @@ public class AppPlatformMetroExtensionsPluginComponentRegistrar : CompilerPlugin
 
   override fun ExtensionStorage.registerExtensions(configuration: CompilerConfiguration) {
     FirExtensionRegistrarAdapter.registerExtension(AppPlatformMetroExtensionsPluginRegistrar())
-    IrGenerationExtension.registerExtension(
-      AppPlatformIrGenerationExtension(
-        generateDeclarationsInIr =
-          AppPlatformCompilerConfiguration.generateClassesInIr(configuration)
-      )
-    )
+    IrGenerationExtension.registerExtension(ContributesRendererIrExtension())
+    IrGenerationExtension.registerExtension(ContributesRobotIrExtension())
+    IrGenerationExtension.registerExtension(ContributesScopedIrExtension())
   }
 }