+++ metro-extensions/contribute/impl-compiler-plugin/src/main/kotlin/software/ralf/app/platform/metro/compiler/scoped/ContributesScopedMetroExtension.kt
@@ -71,8 +71,7 @@ public class ContributesScopedMetroExtension(private val session: FirSession) :
       session: FirSession,
       options: MetroOptions,
       compatContext: CompatContext,
-    ): MetroContributionExtension? {
-      if (options.generateClassesInIr) return null
+    ): MetroContributionExtension {
       return ContributesScopedMetroExtension(session)
     }
   }