+++ metro-extensions/contribute/impl-compiler-plugin/src/test/kotlin/software/ralf/app/platform/metro/compiler/services/MetroRuntimeProvider.kt
@@ -1,6 +1,5 @@
 package software.ralf.app.platform.metro.compiler.services
 
-import dev.zacsweers.metro.compiler.MetroCommandLineProcessor
 import java.io.File
 import org.jetbrains.kotlin.cli.jvm.config.addJvmClasspathRoots
 import org.jetbrains.kotlin.config.CompilerConfiguration
@@ -9,7 +8,6 @@ import org.jetbrains.kotlin.test.model.TestModule
 import org.jetbrains.kotlin.test.services.EnvironmentConfigurator
 import org.jetbrains.kotlin.test.services.RuntimeClasspathProvider
 import org.jetbrains.kotlin.test.services.TestServices
-import software.ralf.app.platform.metro.compiler.AppPlatformCommandLineProcessor
 
 private val metroRuntimeClasspath: List<File> =
   System.getProperty("metroRuntime.classpath")
@@ -22,55 +20,13 @@ fun TestConfigurationBuilder.configureMetroRuntime() {
   useCustomRuntimeClasspathProviders(::MetroRuntimeClasspathProvider)
 }
 
-fun TestConfigurationBuilder.configurePureIrMetroRuntime() {
-  useConfigurators(::PureIrMetroRuntimeEnvironmentConfigurator)
-  useCustomRuntimeClasspathProviders(::MetroRuntimeClasspathProvider)
-}
-
 private class MetroRuntimeEnvironmentConfigurator(testServices: TestServices) :
   EnvironmentConfigurator(testServices) {
   override fun configureCompilerConfiguration(
     configuration: CompilerConfiguration,
     module: TestModule,
   ) {
-    configuration.configureMetro(generateClassesInIr = true, generateContributionHintsInFir = true)
-  }
-}
-
-private class PureIrMetroRuntimeEnvironmentConfigurator(testServices: TestServices) :
-  EnvironmentConfigurator(testServices) {
-  override fun configureCompilerConfiguration(
-    configuration: CompilerConfiguration,
-    module: TestModule,
-  ) {
-    configuration.configureMetro(
-      generateClassesInIr = true,
-      generateContributionHintsInFir = false,
-    )
-  }
-}
-
-private fun CompilerConfiguration.configureMetro(
-  generateClassesInIr: Boolean,
-  generateContributionHintsInFir: Boolean,
-) {
-  addJvmClasspathRoots(metroRuntimeClasspath)
-  MetroCommandLineProcessor().apply {
-    mapOf(
-        "generate-classes-in-ir" to generateClassesInIr,
-        "generate-contribution-hints-in-fir" to generateContributionHintsInFir,
-      )
-      .forEach { (optionName, value) ->
-        val option = pluginOptions.single { it.optionName == optionName }
-        processOption(option, value.toString(), this@configureMetro)
-      }
-  }
-  AppPlatformCommandLineProcessor().apply {
-    processOption(
-      AppPlatformCommandLineProcessor.GENERATE_CLASSES_IN_IR,
-      generateClassesInIr.toString(),
-      this@configureMetro,
-    )
+    configuration.addJvmClasspathRoots(metroRuntimeClasspath)
   }
 }