+++ metro-extensions/contribute/impl-compiler-plugin/src/test/resources/box/contributesrobot/constructorParametersWithoutInject.kt
@@ -1,5 +1,6 @@
 package com.test
 
+import dev.zacsweers.metro.BindingContainer
 import software.ralf.app.platform.inject.robot.ContributesRobot
 import software.ralf.app.platform.robot.Robot
 import software.ralf.app.platform.robot.RobotGraph
@@ -19,6 +20,24 @@ interface MyGraph : RobotGraph {
 }
 
 fun box(): String {
+  if (
+    TestRobot.RobotContribution::class.java.getAnnotation(BindingContainer::class.java) == null
+  ) {
+    return "FAIL: expected RobotContribution to be a BindingContainer"
+  }
+  if (
+    TestRobot.RobotContribution::class.java.declaredMethods.any { it.name == "provideTestRobot" }
+  ) {
+    return "FAIL: expected provider to be moved off the RobotContribution interface"
+  }
+  if (
+    TestRobot.RobotContribution.Companion::class.java.declaredMethods.none {
+      it.name == "provideTestRobot"
+    }
+  ) {
+    return "FAIL: expected provider on RobotContribution companion"
+  }
+
   val graph = createGraph<MyGraph>()
   val robotFactory = graph.robots.getValue(TestRobot::class)
   val robot = robotFactory()